feat: added elastic filter and finalized elastic filter completely. refactored the place of elasticsearch
