+++ langchain/vectorstores/__init__.py
@@ -1,6 +1,6 @@
 """Wrappers on top of vector stores."""
 from langchain.vectorstores.base import VectorStore
-from langchain.vectorstores.elastic_vector_search import ElasticVectorSearch
+from langchain.vectorstores.elasticsearch.elastic_vector_search import ElasticVectorSearch
 from langchain.vectorstores.faiss import FAISS
 from langchain.vectorstores.pinecone import Pinecone
 from langchain.vectorstores.qdrant import Qdrant
