+++ langchain/vectorstores/filters/elastic_filter.py
@@ -1,3 +1,4 @@
+import json
 from typing import List, Any
 
 from langchain.vectorstores.filters.base import VectorStoreFilter
@@ -9,13 +10,14 @@ class ElasticFilter(VectorStoreFilter):
     def __init__(self):
         self.bool_queries = []
 
-    def add_filter_exact_match(self, field_name: str, field_value: List[Any]):
-        should_clause = []
-        for field_val in field_value:
-            should_clause.append({"term": {field_name: field_val}})
-        self.bool_queries.append({"should": should_clause})
+    def add_filter_exact_match(self, field_name: str, field_value: Any):
+        self.bool_queries.append({"term": {"metadata." + field_name: field_value}})
 
-    def to_query_string(self) -> str:
-        return self.query_string
-
-
+    def to_query_string(self, field_prefix: str = None) -> str:
+        return json.dumps(
+            {
+                "bool": {
+                    "must": self.bool_queries
+                }
+            }
+        )
